Refactor config UI forms and widgets
